refactor: move adjustLoan to its own file
