+++ src/game-logic/reducers/index.test.js
@@ -1,5 +1,4 @@
 import { testCrop } from '../../test-utils'
-import { LOAN_INCREASED, LOAN_PAYOFF } from '../../templates'
 import {
   MAX_LATEST_PEER_MESSAGES,
   MAX_PENDING_PEER_MESSAGES,
@@ -22,61 +21,6 @@ jest.mock('../../constants', () => ({
   CROW_CHANCE: 0,
   PRECIPITATION_CHANCE: 0,
 }))
-
-describe('adjustLoan', () => {
-  test('updates state', () => {
-    expect(
-      fn.adjustLoan(
-        { money: 100, loanBalance: 50, todaysNotifications: [] },
-        -25
-      )
-    ).toEqual({
-      money: 75,
-      loanBalance: 25,
-      todaysNotifications: [],
-    })
-  })
-
-  describe('loan payoff', () => {
-    test('shows appropriate notification', () => {
-      const { loansTakenOut, todaysNotifications } = fn.adjustLoan(
-        {
-          money: 100,
-          loanBalance: 50,
-          loansTakenOut: 1,
-          todaysNotifications: [],
-        },
-        -50
-      )
-
-      expect(todaysNotifications).toEqual([
-        { message: LOAN_PAYOFF``, severity: 'success' },
-      ])
-
-      expect(loansTakenOut).toEqual(1)
-    })
-  })
-
-  describe('loan increase', () => {
-    test('shows appropriate notification, updates state', () => {
-      const { loansTakenOut, todaysNotifications } = fn.adjustLoan(
-        {
-          money: 100,
-          loanBalance: 50,
-          todaysNotifications: [],
-          loansTakenOut: 1,
-        },
-        50
-      )
-
-      expect(todaysNotifications).toEqual([
-        { message: LOAN_INCREASED`${100}`, severity: 'info' },
-      ])
-
-      expect(loansTakenOut).toEqual(2)
-    })
-  })
-})
 
 describe('forRange', () => {
   test('calls given reducer on range of plots', () => {
